User: Create user route
