+++ db/books.json
@@ -1,1 +1,25 @@
-[{"title":"The Design of Everyday Things","author":"Don Norman","year":2014,"id":1},{"title":"Eloquent JavaScript, 3rd Edition: A Modern Introduction to Programming","author":"Marijn Haverbeke","year":2001,"id":2},{"title":"Algorithms to Live By","author":"Brian Christian and Thomas L. Griffiths","year":2016,"id":3},{"title":"Things fall apart","author":"Chinua Achebe","year":1958,"id":6},{"title":"A Book","author":"An Author","year":2022,"id":7},{"title":"A Book","author":"An Author","year":2022,"id":8},{"title":"A Book","author":"An Author","year":2022,"id":9},{"title":"A Book","author":"An Author","year":2022,"id":10},{"title":"A Book","author":"An Author","year":2022,"id":11},{"title":"A Book","author":"An Author","year":2022,"id":12},{"title":"A Book","author":"An Author","year":2022,"id":13},{"title":"A Book","author":"An Author","year":2022,"id":14},{"title":"A Book","author":"An Author","year":2022,"id":15},{"title":"A Book","author":"An Author","year":2022,"id":16},{"title":"A Book","author":"An Author","year":2022,"id":17},{"title":"A Book","author":"An Author","year":2022,"id":18},{"title":"A Book","author":"An Author","year":2022,"id":19}]
+[
+	{
+		"title": "The Design of Everyday Things",
+		"author": "Don Norman",
+		"year": 2014,
+		"id": 1
+	},
+	{
+		"title": "Eloquent JavaScript, 3rd Edition: A Modern Introduction to Programming",
+		"author": "Marijn Haverbeke",
+		"year": 2001,
+		"id": 2
+	},
+	{
+		"title": "Algorithms to Live By",
+		"author": "Brian Christian and Thomas L. Griffiths",
+		"year": 2016,
+		"id": 3
+	},
+	{
+		"title": "Things fall apart",
+		"author": "Chinua Achebe",
+		"year": 1958,
+		"id": 4
+	}]
