+++ db/users.json
@@ -1,0 +1,20 @@
+[
+	{
+		"username": "John42",
+		"email": "john123@mail.com",
+		"password": "mypassword1",
+		"role": "admin"
+	},
+	{
+		"username": "Yemi4life",
+		"email": "yemi4@mail.com",
+		"password": "yemipassword2",
+		"role": "reader"
+	},
+	{
+		"username": "Rising234",
+		"email": "rising234@mail.com",
+		"password": "mypassword2",
+		"role": "reader"
+	}
+]
